+++ src/fpjs-provider.tsx
@@ -1,6 +1,7 @@
 import { PropsWithChildren, useCallback, useEffect, useMemo, useRef, useState } from 'react'
 import FpjsContext from './fpjs-context'
 import { FpjsClient, FpjsClientOptions, Agent } from '@fingerprintjs/fingerprintjs-pro-spa'
+import * as packageInfo from '../package.json'
 
 interface FpjsProviderOptions extends FpjsClientOptions {
   /**
@@ -35,7 +36,16 @@ export function FpjsProvider<TExtended extends boolean>({
   loadOptions,
 }: PropsWithChildren<FpjsProviderOptions>) {
   const clientOptions = useMemo(() => {
-    return { cache, cacheTimeInSeconds, cachePrefix, cacheLocation, loadOptions }
+    return {
+      cache,
+      cacheTimeInSeconds,
+      cachePrefix,
+      cacheLocation,
+      loadOptions: {
+        ...loadOptions,
+        integrationInfo: [...(loadOptions.integrationInfo || []), `fingerprintjs-pro-react/${packageInfo.version}`],
+      },
+    }
   }, [cache, cacheTimeInSeconds, cachePrefix, cacheLocation, loadOptions])
 
   const [client, setClient] = useState<FpjsClient>(() => new FpjsClient(clientOptions))
@@ -46,10 +56,15 @@ export function FpjsProvider<TExtended extends boolean>({
     }
   }, [forceRebuild, clientOptions])
 
-  const clientPromise = useRef<Promise<Agent>>()
+  const clientPromise = useRef<Promise<Agent>>(client.init())
+  const firstRender = useRef(true)
 
   useEffect(() => {
-    clientPromise.current = client.init()
+    if (firstRender) {
+      firstRender.current = false
+    } else {
+      clientPromise.current = client.init()
+    }
   }, [client])
 
   const getVisitorData = useCallback(