+++ src/use-visitor-data.ts
@@ -1,6 +1,6 @@
 import FpjsContext, { FpjsContextInterface, QueryResult, VisitorQueryContext } from './fpjs-context'
 import { useCallback, useContext, useEffect, useMemo, useState } from 'react'
-import { VisitorData, GetOptions } from '@fingerprintjs/fingerprintjs-pro-spa'
+import { GetOptions, VisitorData } from '@fingerprintjs/fingerprintjs-pro-spa'
 
 /**
  * ```js
@@ -9,25 +9,26 @@ import { VisitorData, GetOptions } from '@fingerprintjs/fingerprintjs-pro-spa'
  *  *   data,
  *  *   isLoading,
  *  *   error,
- *  *   // A method to be called manually when the `immediate` flag is set to `false`:
+ *  *   // A method to be called manually when the `immediate` field in the config is set to `false`:
  *  *   getData,
- *  * } = useVisitorData({ extended: true }, false);
+ *  * } = useVisitorData({ extended: true }, { immediate: false });
  *  * ```
  *  *
  *  * Use the `useVisitorData` hook in your components to perform identification requests with the FingerprintJS API.
  *
- * @param config options for the `fp.get()` request
- * @param immediate determines whether the `getData()` method will be called immediately after the hook mounts or not,
+ * @param getOptions options for the `fp.get()` request
+ * @param config config for the hook
  */
 export function useVisitorData<TExtended extends boolean>(
-  config: GetOptions<TExtended> = {},
-  immediate = true
+  getOptions: GetOptions<TExtended> = {},
+  config: UseVisitorDataConfig = defaultUseVisitorDataConfig
 ): VisitorQueryContext<TExtended> {
-  const { extendedResult, timeout, tag, linkedId } = config ?? {}
-  const getOptions = useMemo(
+  const { extendedResult, timeout, tag, linkedId } = getOptions ?? {}
+  const memoizedOptions = useMemo(
     () => ({ extendedResult, timeout, tag, linkedId }),
     [extendedResult, timeout, tag, linkedId]
   )
+  const { immediate } = config
   const { getVisitorData } = useContext<FpjsContextInterface<TExtended>>(FpjsContext)
 
   const initialState = { isLoading: false }
@@ -38,7 +39,7 @@ export function useVisitorData<TExtended extends boolean>(
       try {
         setState((state) => ({ ...state, isLoading: true }))
 
-        const result = await getVisitorData(getOptions, ignoreCache)
+        const result = await getVisitorData(memoizedOptions, ignoreCache)
         setState((state) => ({ ...state, data: result, isLoading: false, error: undefined }))
         return result
       } catch (error) {
@@ -51,7 +52,7 @@ export function useVisitorData<TExtended extends boolean>(
         setState((state) => (state.isLoading ? { ...state, isLoading: false } : state))
       }
     },
-    [getOptions, getVisitorData]
+    [memoizedOptions, getVisitorData]
   )
 
   useEffect(() => {
@@ -69,3 +70,12 @@ export function useVisitorData<TExtended extends boolean>(
     error,
   }
 }
+
+export interface UseVisitorDataConfig {
+  /**
+   * Determines whether the `getData()` method will be called immediately after the hook mounts or not
+   */
+  immediate: boolean
+}
+
+const defaultUseVisitorDataConfig = { immediate: true }