+++ CHANGELOG.md
@@ -1,6 +1,31 @@
-## 0.2.1
-### Chores
+### 0.3.0 (2022-03-14)
+
+##### Chores
+
+*  added source maps ([f922344e](https://github.com/fingerprintjs/fingerprintjs-pro-react/commit/f922344e47fc04a8bbce0f10cc99dbc4e5945c20))
+
+##### New Features
+
+*  pass config to the useVisitorData hook ([750733a3](https://github.com/fingerprintjs/fingerprintjs-pro-react/commit/750733a39de8cf30db1be5576f30c618c5bdb060))
+
+##### Bug Fixes
+
+*  race condition (getData called before init) ([0c787786](https://github.com/fingerprintjs/fingerprintjs-pro-react/commit/0c78778652b598b329d3056ec002bf85d0733460))
+
+#### 0.2.2 (2022-03-11)
+
+##### Chores
+
+*  add automatic changelog generation ([9126a8e6](https://github.com/fingerprintjs/fingerprintjs-pro-react/commit/9126a8e6a7f6b83eb11f79cf47f9312e49f5cff2))
+*  bumped fingerprintjs-pro-spa version ([4c5237b8](https://github.com/fingerprintjs/fingerprintjs-pro-react/commit/4c5237b87b6f9bcb1b95ab49b0eabb12f62c2a3a))
+
+##### Bug Fixes
+
+*  eslint configs ([1a9c57e0](https://github.com/fingerprintjs/fingerprintjs-pro-react/commit/1a9c57e059977de3c8b31a3a1fe2ec535a6f093d))
+
+### 0.2.1
+##### Chores
 * Bumped `fingerprintjs-spa-pro` version
 
-## 0.2.0
+### 0.2.0
 * Initial release